+++ src/components/pages/hospital/Certified/SubmitDocument/type.ts
@@ -1,3 +1,3 @@
-import { AuthStatus } from "types/admin.type";
-
-export type SubmitDocumentProps = { requestStatus?: AuthStatus };
+export type SubmitDocumentProps = {
+  requestStatus?: boolean;
+};
