+++ src/utils/apis/admin.ts
@@ -1,11 +1,13 @@
 import {
   ImageReqType,
   InfoResType,
+  PatchHospitalAuthReqType,
   SignInReqType,
   SignInResType,
   SignUpReqType,
 } from "types/admin.type";
 import { AuthInstance, DefaultInstance } from ".";
+import { ApproveDetailResType, ListResType } from "types/hospitalAuth";
 
 const router = "/admin";
 
@@ -41,4 +43,32 @@ export const GetUserInfoApi = async (): Promise<InfoResType> => {
 /** 계정 삭제 */
 export const DeleteAccountApi = async () => {
   await AuthInstance.delete(`${router}/secession`);
+};
+
+/** 병원관계자 승인 리스트 목록 */
+export const GetHospitalListApi = async (): Promise<ListResType> => {
+  const { data }: { data: ListResType } = await AuthInstance.get(
+    `${router}/list`
+  );
+  return data;
+};
+
+/** 병원관계자 승인 상세보기 - (hospitalId) */
+export const GetHospitalDetailApi = async (
+  hospitalId: number
+): Promise<ApproveDetailResType> => {
+  const { data }: { data: ApproveDetailResType } = await AuthInstance.get(
+    `${router}/${hospitalId}`
+  );
+  return data;
+};
+
+/** 병원 관계자 승인 */
+export const PatchHospitalAuth = async ({
+  id,
+  auth_status,
+}: PatchHospitalAuthReqType) => {
+  await AuthInstance.patch(`${router}/auth/${id}`, {
+    auth_status,
+  });
 };
